requirements.txt
# LangGraph
langgraph>=0.2.0
langgraph-checkpoint>=1.0.0

# LangChain
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0

# OpenAI
openai>=1.0.0

# Embedding / Vector DB
sentence-transformers>=2.2.0
chromadb>=0.4.0

# PDF
pypdf>=3.0.0

# Web UI
gradio>=4.0.0

# HTTP Requests (Google Search API)
requests>=2.28.0

# Utils
pyyaml>=6.0
python-dotenv>=1.0.0

#in terminal : pip install -r requirements.txt